+++ src/Modules/Store/StoreMaintenanceTransactions.php
@@ -2,13 +2,11 @@
 
 namespace Foodsharing\Modules\Store;
 
-use DateInterval;
-use DateTime;
+use Carbon\Carbon;
 use Foodsharing\Modules\Core\DBConstants\Foodsaver\UserOptionType;
 use Foodsharing\Modules\Core\DBConstants\Store\CooperationStatus;
 use Foodsharing\Modules\Foodsaver\FoodsaverGateway;
 use Foodsharing\Modules\Settings\SettingsGateway;
-use Foodsharing\Modules\Store\DTO\PickupInformation;
 use Foodsharing\Utility\EmailHelper;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
@@ -16,75 +14,78 @@ class StoreMaintenanceTransactions
 {
     public function __construct(
         private readonly StoreGateway $storeGateway,
-        private readonly PickupTransactions $pickupTransactions,
         private readonly EmailHelper $emailHelper,
         private readonly TranslatorInterface $translator,
         private readonly FoodsaverGateway $foodsaverGateway,
         private readonly SettingsGateway $settingsGateway,
+        private readonly PickupGateway $pickupGateway,
     ) {
     }
 
     public function triggerFetchWarningNotification(): array
     {
-        $activeStores = $this->storeGateway->getAllStores(
-            [CooperationStatus::COOPERATION_ESTABLISHED]
-        );
+        $activeStores = $this->storeGateway->getAllStores([CooperationStatus::COOPERATION_ESTABLISHED]);
 
-        $start = new DateTime(); // Now
-        $end = (new DateTime())->add(DateInterval::createFromDateString('48 hours')); // 48 hours later
-
-        $foodsavers = [];
+        $start = Carbon::now();
+        $end = $start->copy()->addDays(2);
+        $fetchWarnings = [];
         $storesWithNotification = 0;
-        $totalCountPickups = 0;
-        $totalCountEmptyPickups = 0;
         $mailsDisabledViaSettings = 0;
 
+        // collect empty slot information:
         foreach ($activeStores as $store) {
-            $allPickups = $this->pickupTransactions->getPickupsWithUsersForPickupsInRange($store['id'], $start, $end);
-            $totalCountPickups += count($allPickups);
+            $pickups = $this->pickupGateway->getPickupSlots($store['id'], $start, $end, $end);
+            $occupiedSlots = array_sum(array_map('count', array_column($pickups, 'occupiedSlots')));
+            $emptySlots = array_sum(array_column($pickups, 'totalSlots')) - $occupiedSlots;
 
-            $emptyPickups = array_filter(
-                $allPickups,
-                fn (PickupInformation $pickup) => !$pickup->hasConfirmedUser()
-            );
-
-            $countEmptyPickups = count($emptyPickups);
-            $totalCountEmptyPickups += $countEmptyPickups;
-
-            if ($countEmptyPickups > 0) {
+            if ($emptySlots > 0) {
                 ++$storesWithNotification;
-
                 $storeManagers = $this->storeGateway->getStoreManagers($store['id']);
-                $managaerMailSettings = $this->settingsGateway->getUsersOption($storeManagers, UserOptionType::DISABLE_PICKUP_REMINDER);
-                foreach ($managaerMailSettings as $mailsSetting) {
-                    if ($mailsSetting['option']) {
-                        ++$mailsDisabledViaSettings;
-                        continue;
+                foreach ($storeManagers as $storeManagerId) {
+                    if (!isset($fetchWarnings[$storeManagerId])) {
+                        $storeManager = $this->foodsaverGateway->getFoodsaver($storeManagerId);
+                        $fetchWarnings[$storeManagerId] = [
+                            'email' => $storeManager['email'],
+                            'anrede' => $this->translator->trans('salutation.' . $storeManager['geschlecht']),
+                            'name' => $storeManager['name'],
+                            'stores' => [],
+                        ];
                     }
-                    $foodsaverId = $mailsSetting['userId'];
-                    $foodsavers[] = $foodsaverId;
-
-                    $fs = $this->foodsaverGateway->getFoodsaver($foodsaverId);
-                    $this->emailHelper->tplMail('chat/fetch_warning', $fs['email'], [
-                        'anrede' => $this->translator->trans('salutation.' . $fs['geschlecht']),
-                        'name' => $fs['name'],
-                        'betrieb' => $store['name'],
-                        'link' => BASE_URL . '/?page=fsbetrieb&id=' . $store['id']
-                    ]);
+                    $fetchWarnings[$storeManagerId]['stores'][] = [
+                        'link' => BASE_URL . '/store/' . $store['id'],
+                        'name' => $store['name'],
+                        'count' => $emptySlots
+                    ];
                 }
             }
+        }
+
+        // Remove mails for users with disabled pickup reminder:
+        $storeManagerIds = array_keys($fetchWarnings);
+        $batchedIds = array_chunk($storeManagerIds, 100);
+        foreach ($batchedIds as $batch) {
+            $disabledSettings = $this->settingsGateway->getUsersOption($batch, UserOptionType::DISABLE_PICKUP_REMINDER);
+            foreach ($disabledSettings as $setting) {
+                if ($setting['option']) {
+                    unset($fetchWarnings[$setting['userId']]);
+                    ++$mailsDisabledViaSettings;
+                }
+            }
+        }
+
+        // Send mails:
+        foreach ($fetchWarnings as $storeManagerId => $fetchWarning) {
+            $this->emailHelper->tplMail('chat/fetch_warning', $fetchWarning['email'],
+                $fetchWarning + ['settings' => BASE_URL . '/user/current/settings?sub=info']);
         }
 
         return [
             'start' => $start->format('c'),
             'end' => $end->format('c'),
-            'count_stores' => count($activeStores),
-            'count_stores_with_notifications' => $storesWithNotification,
-            'count_unique_foodsavers' => count(array_unique($foodsavers)),
-            'count_warned_foodsavers' => count($foodsavers),
-            'count_total_pickups' => $totalCountPickups,
-            'count_total_empty_pickups' => $totalCountEmptyPickups,
-            'mails_disabled_via_settings' => $mailsDisabledViaSettings,
+            'stores checked' => count($activeStores),
+            'stores with notifications' => $storesWithNotification,
+            'mails disabled via settings' => $mailsDisabledViaSettings,
+            'warned foodsavers' => count($fetchWarnings),
         ];
     }
 }